requirements.txt
Flask==3.1.0
gunicorn==23.0.0
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
greenlet==3.1.1
